fix(observability): re-point kb-sync Sentry alert to kb-sync.unexpected after tenant-mint convergence (#5005)
The kb/sync resolver convergence removed the last kb-sync.tenant-mint emit
site, which was the surviving slug keeping kb_tenant_mint_silent_fallback
armed. Re-point the alert to the route's surviving silent-failure surface
(top-level catch, op=kb-sync.unexpected, feature=kb-route-helpers) and rename
the resource to kb_sync_silent_failure. Rewrite the canonical route test
(test/server/kb-sync-route.test.ts) for the resolver path and drop the
duplicate test/kb-sync-route.test.ts.
